requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
db-dtypes==1.4.3
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
fastuuid==0.14.0
filelock==3.20.3
frozenlist==1.7.0
fsspec==2026.1.0
fuzzywuzzy==0.18.0
google-api-core==2.25.1
google-auth==2.47.0
google-auth-oauthlib==1.2.4
google-cloud-bigquery==3.35.1
google-cloud-bigquery-storage==2.36.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-genai==1.58.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface_hub==1.3.2
idna==3.10
importlib_metadata==8.7.1
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
litellm==1.80.16
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.3
numpy==2.3.2
oauthlib==3.3.1
openai==2.15.0
packaging==25.0
pandas==2.3.1
pandas-gbq==0.33.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.2.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.12.0
pydantic_core==2.33.2
pydata-google-auth==1.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.37.0
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.2
rpds-py==0.30.0
rsa==4.9.1
sentry-sdk==2.33.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
strsimpy==0.2.1
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.22.2
tqdm==4.67.1
typer==0.16.0
typer-slim==0.21.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
